Switch to using floats and fix debug logging

+++ src/microbe_stage/ISpawned.cs
@@ -10,7 +10,7 @@ public interface ISpawned : IEntity
     int DespawnRadiusSquared { get; set; }
 
     /// <summary>
-    ///   How much this entity contributes to the entity limit (based on Node complexity)
+    ///   How much this entity contributes to the entity limit relative to a single node
     /// </summary>
-    int EntityWeight { get; }
+    float EntityWeight { get; }
 }
